requirements.txt
altgraph==0.17.4
asgiref==3.8.1
auto-py-to-exe==2.46.0
beautifulsoup4==4.13.4
bottle==0.13.3
bottle-websocket==0.2.9
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
distlib==0.3.9
Django==4.2.22
django-bootstrap-v5==1.0.11
django-bootstrap4==25.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Eel==0.18.1
filelock==3.18.0
future==1.0.0
gevent==25.4.2
gevent-websocket==0.10.1
greenlet==3.2.2
idna==3.10
Markdown==3.8
numpy==2.3.0
package_name==0.1
packaging==25.0
pandas==2.3.0
pefile==2023.2.7
pillow==11.2.1
platformdirs==4.3.8
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==80.4.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.1
whitenoise==6.6.0
zope.event==5.0
zope.interface==7.2
